Splitting builds and tests
This requires building in predefined directories (not specific to each
runner).

Also, we are not using artifacts, so there is no dependency
relationship in that sense between build and test jobs.
However, GitLab introduced a new dependency relationship between jobs,
allowing to start a job in a later stage if all its dependencies have
run. (see "needs:" keyword)

Prevent the job from cloning when not necessary, setting GIT_STRATEGY
to "none".