Moving back cudatookit definition:
The definition of cudatoolkit is moved back to project-specific
host-configs files. In fact, ideally we would like to share
the cudatoolkit configuration, but the combination "compiler+cuda"
should remains project specific. In order to avoid direct referencing
from "project host-configs" to ".gitlab host-configs" we keep cuda
config on the project side.