Update CI with new compiler list
GitLab CI now uses a new set of compilers for both toss3 and blueos.
These compilers were selected via discussion with users to ensure that
we are testing against the same compiler vendors and major.minor
versions.

The compiler configurations are provided by the
scripts/radiuss-spack-configs submodule, and the jobs are in the
.gitlab/ directory.