Force gcc to treate Replay tool object as POD
By type casting the pointer to an AllocatorTableEntry to be 'void*',
this will circumvent the typecheck being done by gcc 8 or higher to
determine whether the object is a POD or not. (see:
https://github.com/dotnet/runtime/issues/12684)

Prior to this change, this warning would be triggered in GCC 10.2.1:
  error: 'void* memset(void*, int, size_t)' clearing an object of
  non-trivial type`

It is unclear why gcc is considering this POD to be non-trivial.