Use blt namespace for hip targets (#692)
The cmake configuration system of Umpire was changed to use blt::blt_runtime and blt::blt for HOST runtime and DEVICE code respectively. To support this, the docker CI instance for HIP was updated to use ROCM 4.3.1.

CI for hip was modified to stop treating compiler warnings as errors since the ROCM 4.3.1 cmake configuration file is adding a command line option to the compiler which is not supported yet causing the compiler to generate a warning that it is ignoring the option.