Add fast path for allocate events
This still needs work, but the performance improvements are there.  I
want to review these changes because the raw interface still needs to be
pushed down in to the various specific store objects (it is currently
all being done at the event level and in a very json-specific way which
is wrong).

We just need to determine the best ways of accomplishing this.