Refactor allocate event generation
Phase 1

This is the first part of the refactoring effort which simply involves
changing the highest level calls to the event API from the allocate and
deallocate functions.  The next phase will be to push the json
generation code down from the event layer in to its proper place in the
json store area.