Add ASAN CI to azure pipeline (#707)
CI tests with ASAN have been enabled for azure pipelines with the exception of the operation_tests which were causing an OOM due to the default container resource limit and the size of the test. CI tests for the operation_tests are still done via gitlab CI which runs on physical LC compute nodes.