Fix `find_first_set()` on Windows
This corrects the usage of `_BitScanForward()` in the Windows
implementation of `find_first_set()`. `_BitScanForward()` uses its
return value (type: `unsigned char`) to indicate if a set bit was found.
If a set bit was found, then the parameter `index` stores the position
of that bit (zero-based indexing).

On Linux, `ffs()` uses its return value (type: `int`) to store the
position of the first set bit found (one-based indexing); a return value
of 0 indicates that no bits are set.