Adding umpire_util dependency to umpire_device lib (#735)
In order to allow error handling for the DeviceAllocator, we need to add the umpire_util dependency to the umpire_device lib. There were some other additional edits that were made so that error handling in the DeviceAllocator files were correct. We also added "+device_alloc" to a lassen CI job so that the DeviceAllocator is tested in the CI and these issues can be found sooner.