Add UMAP allocation resource (#711)
This adds UMAP as a memory resource for Umpire. When enabled via '-DUMPIRE_ENABLE_UMAP=On', the umpire configuration will look for an installation of umap pointed to by -DUMAP_ROOT=path. When built, the the FileMemory object will be configured to use UMAP instead of MMAP for its allocations.

For CI, build-only testing is current all that is supported. Further testing will be enabled when TOSS4 rolls out and we have full read/write access to the userfaultfd capabilities in Linux.

Co-authored-by: Abhik Sarkar <sarkar6@llnl.gov>