cmake: properly add Fortran module directories
The fix in the INSTALL_INTERFACE is necessary so that a project that
does

```
find_package(umpire)

[...]
target_link_libraries(my_fortran_exe PRIVATE umpire)
```

actually will find the `umpire_mod.mod` file, which gets installed
into `include/umpire/`.

The change in the BUILD_INTERFACE is for the case where umpire is used in
another cmake project via `add_subdirectory()`, which is probably not an
offically supported case, but it can be quite useful.