BUGFIX: properly set ID for device allocator
The ``umpire::resource::resource_to_device_id()`` method has logic
to detect the device ID based on the name of the memory resource.
However, if the calling code has set a particular GPU to the
process, e.g., via calling ``cudaSetDevice()`` or ``hipSetDevice()``,
the default device allocator that Umpire provides may have the
wrong device ID. Consequently, all memory allocations/operations
would be directed to the wrong GPU on the system.

This commit fixes this issue by adding logic to query the CUDA or HIP
runtime for the correct device ID to use if the caller has not
encoded the device ID to the name of the memory resource.

Resolves #798