sycl: support 2023 oneapi release
The CL/sycl.hpp header has been removed in the 2023 release.
Note that the SYCL 2020 standard prescribes sycl/sycl.hpp as location,
so this change in oneapi brings it in accordinace with the standard.

dpcpp has also been depreciated, using icpx with -fsycl is now
preferred and dcpp will be removed. This attempts to preserve support
for older oneapi releases via the sycl_compat.hpp header.