Use getActualHighwaterMark instead of getHighWatermark to determine coalesced pool size
The non-actual high water mark does not include the overallocated size
when rounding user allocations up to the alignment. This means that the
non-actual high water mark may be too small when repeatedly allocating
and deallocating the same sizes in a pool, leading to unnecessary
reallocation of the backing buffers.